bugfix
in bang(), there was a problem because activation was set to 0.0 after sending midi events. this resulted in huge thread overload and error when the amount of threads was not limited to 10 or so. this was because activation did not set to 0.0 and the bang-thread was called as activation rose above the threshold, which is not the purpose.